Splitted HDL ATMegaxx0_1 TIMER into several files
